+++ CHANGELOG.md
@@ -11,6 +11,7 @@ The official release will have a defined and more stable API. If you are already
     * `lambdacd.presentation.pipeline-state/overall-build-status`
     * `lambdacd.presentation.pipeline-state/latest-most-recent-update` 
     * `lambdacd.presentation.pipeline-state/earliest-first-update` 
+    * `lambdacd.presentation.pipeline-state/build-duration` 
   * UI: Link LambdaCD header to "/" to link back to overview in cases with multiple pipelines (#82)
 * Bugs: 
   * `with-workspace` now creates temporary directories in the home-dir (#79)